fix build errors in GH actions by moving rollup-plugin-serve and watch into separate script (npm run dev)
